requirements.txt
cycler==0.10.0
gitdb==4.0.7
GitPython==3.1.17
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
smmap==4.0.0
